# swagger.yaml
openapi: 3.0.1
info:
  title: NFV Test API
  description: Test api for client side network operations
  version: '1.0'
servers:
  - url: /api/v2
...
paths:
...
  /routes:
    get:
      tags:
        - routes
      summary: Get all routes on the host
      operationId: get_all_routes
      responses:
        '200':
          description: Get all routes on the host
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Route'
        '500':
          description: >-
            An error occurred when trying to process the request, this can also
            be because of bad input from the user
          content: {}
    post:
      tags:
        - routes
      summary: Create a route on the host
      description: >-
        The route is identified by its destination address, if another route
        with the same destination

        already exists, a conflict error is raised.
      operationId: post_all_routes
      requestBody:
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/RouteCreate'
        required: true
      responses:
        '201':
          description: A new route has been created
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Route'
        '409':
          description: Another route with the same name already exists
          content: {}
        '422':
          description: Couldn't update route because of invalid input
          content: {}
        '500':
          description: >-
            An error occurred when trying to process the request, this can also
            be because of bad input from the user
          content: {}
      x-codegen-request-body-name: payload
  /routes/ns/{ns_name}:
    get:
      tags:
        - routes
      summary: Get all routes on the host
      operationId: get_all_routes_in_namespace
      parameters:
        - name: ns_name
          in: path
          description: The name of the namespace in which routes belong
          required: true
          schema:
            type: string
      responses:
        '200':
          description: Get all routes on the host
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Route'
        '500':
          description: >-
            An error occurred when trying to process the request, this can also
            be because of bad input from the user
          content: {}
    post:
      tags:
        - routes
      summary: Create a route on the host
      description: >-
        The route is identified by its destination address, if another route
        with the same destination

        already exists, a conflict error is raised.
      operationId: post_all_routes_in_namespace
      parameters:
        - name: ns_name
          in: path
          description: The name of the namespace in which routes belong
          required: true
          schema:
            type: string
      requestBody:
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/RouteCreate'
        required: true
      responses:
        '201':
          description: A new route has been created
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Route'
        '409':
          description: Another route with the same name already exists
          content: {}
        '422':
          description: Couldn't update route because of invalid input
          content: {}
        '500':
          description: >-
            An error occurred when trying to process the request, this can also
            be because of bad input from the user
          content: {}
      x-codegen-request-body-name: payload
  /routes/ns/{ns_name}/{dst_addr}:
    get:
      tags:
        - routes
      summary: Get a route on the host
      description: The route is identified by its destination.
      operationId: get_one_route_in_namespace
      parameters:
        - name: dst_addr
          in: path
          description: >-
            The destination ("default" or any address with a prefix length) of
            the route we mean to select
          required: true
          schema:
            type: string
        - name: ns_name
          in: path
          description: The name of the namespace in which the route belong
          required: true
          schema:
            type: string
      responses:
        '200':
          description: Found an route with a matching destination
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Route'
        '404':
          description: Couldn't find any route with given destination
          content: {}
        '500':
          description: >-
            An error occurred when trying to process the request, this can also
            be because of bad input from the user
          content: {}
    delete:
      tags:
        - routes
      summary: Delete a route from the host
      description: >-
        The route is identified by its destination. This method is idempotent,
        if the route

        doesn't exist it won't try to delete it again, and consider the deletion
        successful.
      operationId: delete_one_route_in_namespace
      parameters:
        - name: dst_addr
          in: path
          description: >-
            The destination ("default" or any address with a prefix length) of
            the route we mean to select
          required: true
          schema:
            type: string
        - name: ns_name
          in: path
          description: The name of the namespace in which the route belong
          required: true
          schema:
            type: string
      responses:
        '200':
          description: The route doesn't exist anymore
          content: {}
        '500':
          description: >-
            An error occurred when trying to process the request, this can also
            be because of bad input from the user
          content: {}
    patch:
      tags:
        - routes
      summary: Update a route on the host
      description: The route is identified by its destination.
      operationId: patch_one_route_in_namespace
      parameters:
        - name: dst_addr
          in: path
          description: >-
            The destination ("default" or any address with a prefix length) of
            the route we mean to select
          required: true
          schema:
            type: string
        - name: ns_name
          in: path
          description: The name of the namespace in which the route belong
          required: true
          schema:
            type: string
      requestBody:
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/RouteUpdate'
        required: true
      responses:
        '200':
          description: The route has been updated
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Route'
        '404':
          description: Couldn't find any route with given destination address
          content: {}
        '422':
          description: Couldn't update route because of invalid input
          content: {}
        '500':
          description: >-
            An error occurred when trying to process the request, this can also
            be because of bad input from the user
          content: {}
      x-codegen-request-body-name: payload
  /routes/ns/{ns_name}/{dst_addr}/{dst_prefix_len}:
    get:
      tags:
        - routes
      summary: Get a route on the host
      description: The route is identified by its destination.
      operationId: get_one_route_with_prefix_in_namespace
      parameters:
        - name: dst_addr
          in: path
          description: The address of the destination of the route
          required: true
          schema:
            type: string
        - name: dst_prefix_len
          in: path
          description: The prefix length of the destination of the route
          required: true
          schema:
            type: integer
        - name: ns_name
          in: path
          description: The name of the namespace in which the route belong
          required: true
          schema:
            type: string
      responses:
        '200':
          description: Found an route with a matching destination
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Route'
        '404':
          description: Couldn't find any route with given destination
          content: {}
        '500':
          description: >-
            An error occurred when trying to process the request, this can also
            be because of bad input from the user
          content: {}
    delete:
      tags:
        - routes
      summary: Delete a route from the host
      description: >-
        The route is identified by its destination. This method is idempotent,
        if the route

        doesn't exist it won't try to delete it again, and consider the deletion
        successful.
      operationId: delete_one_route_with_prefix_in_namespace
      parameters:
        - name: dst_addr
          in: path
          description: The address of the destination of the route
          required: true
          schema:
            type: string
        - name: dst_prefix_len
          in: path
          description: The prefix length of the destination of the route
          required: true
          schema:
            type: integer
        - name: ns_name
          in: path
          description: The name of the namespace in which the route belong
          required: true
          schema:
            type: string
      responses:
        '200':
          description: The route doesn't exist anymore
          content: {}
        '500':
          description: >-
            An error occurred when trying to process the request, this can also
            be because of bad input from the user
          content: {}
    patch:
      tags:
        - routes
      summary: Update a route on the host
      description: The route is identified by its destination.
      operationId: patch_one_route_with_prefix_in_namespace
      parameters:
        - name: dst_addr
          in: path
          description: The address of the destination of the route
          required: true
          schema:
            type: string
        - name: dst_prefix_len
          in: path
          description: The prefix length of the destination of the route
          required: true
          schema:
            type: integer
        - name: ns_name
          in: path
          description: The name of the namespace in which the route belong
          required: true
          schema:
            type: string
      requestBody:
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/RouteUpdate'
        required: true
      responses:
        '200':
          description: The route has been updated
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Route'
        '404':
          description: Couldn't find any route with given destination address
          content: {}
        '422':
          description: Couldn't update route because of invalid input
          content: {}
        '500':
          description: >-
            An error occurred when trying to process the request, this can also
            be because of bad input from the user
          content: {}
      x-codegen-request-body-name: payload
  /routes/{dst_addr}:
    get:
      tags:
        - routes
      summary: Get a route on the host
      description: The route is identified by its destination.
      operationId: get_one_route
      parameters:
        - name: dst_addr
          in: path
          description: >-
            The destination ("default" or any address with a prefix length) of
            the route we mean to select
          required: true
          schema:
            type: string
      responses:
        '200':
          description: Found an route with a matching destination
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Route'
        '404':
          description: Couldn't find any route with given destination
          content: {}
        '500':
          description: >-
            An error occurred when trying to process the request, this can also
            be because of bad input from the user
          content: {}
    delete:
      tags:
        - routes
      summary: Delete a route from the host
      description: >-
        The route is identified by its destination. This method is idempotent,
        if the route

        doesn't exist it won't try to delete it again, and consider the deletion
        successful.
      operationId: delete_one_route
      parameters:
        - name: dst_addr
          in: path
          description: >-
            The destination ("default" or any address with a prefix length) of
            the route we mean to select
          required: true
          schema:
            type: string
      responses:
        '200':
          description: The route doesn't exist anymore
          content: {}
        '500':
          description: >-
            An error occurred when trying to process the request, this can also
            be because of bad input from the user
          content: {}
    patch:
      tags:
        - routes
      summary: Update a route on the host
      description: The route is identified by its destination.
      operationId: patch_one_route
      parameters:
        - name: dst_addr
          in: path
          description: >-
            The destination ("default" or any address with a prefix length) of
            the route we mean to select
          required: true
          schema:
            type: string
      requestBody:
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/RouteUpdate'
        required: true
      responses:
        '200':
          description: The route has been updated
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Route'
        '404':
          description: Couldn't find any route with given destination address
          content: {}
        '422':
          description: Couldn't update route because of invalid input
          content: {}
        '500':
          description: >-
            An error occurred when trying to process the request, this can also
            be because of bad input from the user
          content: {}
      x-codegen-request-body-name: payload
  /routes/{dst_addr}/{dst_prefix_len}:
    get:
      tags:
        - routes
      summary: Get a route on the host
      description: The route is identified by its destination.
      operationId: get_one_route_with_prefix
      parameters:
        - name: dst_addr
          in: path
          description: The address of the destination of the route
          required: true
          schema:
            type: string
        - name: dst_prefix_len
          in: path
          description: The prefix length of the destination of the route
          required: true
          schema:
            type: integer
      responses:
        '200':
          description: Found an route with a matching destination
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Route'
        '404':
          description: Couldn't find any route with given destination
          content: {}
        '500':
          description: >-
            An error occurred when trying to process the request, this can also
            be because of bad input from the user
          content: {}
    delete:
      tags:
        - routes
      summary: Delete a route from the host
      description: >-
        The route is identified by its destination. This method is idempotent,
        if the route

        doesn't exist it won't try to delete it again, and consider the deletion
        successful.
      operationId: delete_one_route_with_prefix
      parameters:
        - name: dst_addr
          in: path
          description: The address of the destination of the route
          required: true
          schema:
            type: string
        - name: dst_prefix_len
          in: path
          description: The prefix length of the destination of the route
          required: true
          schema:
            type: integer
      responses:
        '200':
          description: The route doesn't exist anymore
          content: {}
        '500':
          description: >-
            An error occurred when trying to process the request, this can also
            be because of bad input from the user
          content: {}
    patch:
      tags:
        - routes
      summary: Update a route on the host
      description: The route is identified by its destination.
      operationId: patch_one_route_with_prefix
      parameters:
        - name: dst_addr
          in: path
          description: The address of the destination of the route
          required: true
          schema:
            type: string
        - name: dst_prefix_len
          in: path
          description: The prefix length of the destination of the route
          required: true
          schema:
            type: integer
      requestBody:
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/RouteUpdate'
        required: true
      responses:
        '200':
          description: The route has been updated
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Route'
        '404':
          description: Couldn't find any route with given destination address
          content: {}
        '422':
          description: Couldn't update route because of invalid input
          content: {}
        '500':
          description: >-
            An error occurred when trying to process the request, this can also
            be because of bad input from the user
          content: {}
      x-codegen-request-body-name: payload
...
components:
  schemas:
...
    RouteCreate:
      title: RouteCreate
      description: Input for creating a route
      type: object
      properties:
        dst:
          title: Dst
          anyOf:
            - type: string
              format: ipv4interface
            - type: string
              format: ipv6interface
            - enum:
                - default
              type: string
        gateway:
          title: Gateway
          anyOf:
            - type: string
              format: ipv4
            - type: string
              format: ipv6
        dev:
          title: Dev
          pattern: ^[0-9A-Z-a-z@#$_\-.]{1,16}$
          type: string
      required:
        - dst
        - dev
    Route:
      title: Route
      description: A route as return by the command `ip route`
      type: object
      properties:
        type:
          $ref: '#/components/schemas/Type'
        dst:
          title: Dst
          anyOf:
            - type: string
              format: ipv4interface
            - type: string
              format: ipv6interface
            - enum:
                - default
              type: string
        gateway:
          title: Gateway
          anyOf:
            - type: string
              format: ipv4
            - type: string
              format: ipv6
        dev:
          title: Dev
          pattern: ^[0-9A-Z-a-z@#$_\-.]{1,16}$
          type: string
        protocol:
          title: Protocol
          anyOf:
            - $ref: '#/components/schemas/Protocol'
            - type: integer
            - type: string
              pattern: ^[0-9A-Z-a-z@#$_\-.]{1,16}$
        scope:
          title: Scope
          anyOf:
            - $ref: '#/components/schemas/Scope'
            - type: integer
        flags:
          title: Flags
          type: array
          items:
            type: string
            pattern: ^[0-9A-Z-a-z@#$_\-.]{1,16}$
        pref_src:
          title: Prefsrc
          anyOf:
            - type: string
              format: ipv4
            - type: string
              format: ipv6
        metric:
          title: Metric
          type: integer
      required:
        - type
        - dst
        - dev
        - protocol
        - scope
        - flags
    RouteUpdate:
      title: RouteUpdate
      description: Input for updating a route
      type: object
      properties:
        gateway:
          title: Gateway
          anyOf:
            - type: string
              format: ipv4
            - type: string
              format: ipv6
        dev:
          title: Dev
          pattern: ^[0-9A-Z-a-z@#$_\-.]{1,16}$
          type: string
      required:
        - dev
...
    Scope:
      title: Scope
      description: An enumeration.
      enum:
        - global
        - link
        - host
      type: string
...
    Type:
      title: Type
      description: An enumeration.
      enum:
        - unicast
        - local
        - broadcast
        - multicast
        - throw
        - unreachable
        - prohibit
        - blackhole
        - nat
      type: string
    Protocol:
      title: Protocol
      description: An enumeration.
      enum:
        - boot
        - static
        - kernel
        - redirect
        - ra
      type: string
...